split module rest_api to http + http_server
Why:

* Follow file structure conventions
