+++ wazo_call_logd/plugins/support_center/resource.py
@@ -4,7 +4,7 @@
 from flask import request
 from xivo.auth_verifier import required_acl
 from xivo.tenant_flask_helpers import token, Tenant
-from wazo_call_logd.rest_api import AuthResource
+from wazo_call_logd.http import AuthResource
 
 from .schema import (
     AgentStatisticsListRequestSchema,
